+++ crates/fluvio-sc-schema/src/objects/create.rs
@@ -29,7 +29,7 @@ pub struct CommonCreateRequest {
 
 impl Request for ObjectApiCreateRequest {
     const API_KEY: u16 = AdminPublicApiKey::Create as u16;
-    const DEFAULT_API_VERSION: i16 = 4;
+    const DEFAULT_API_VERSION: i16 = 6;
     type Response = Status;
 }