+++ crates/fluvio-spu/src/services/public/tests/produce.rs
@@ -5,8 +5,9 @@ use dataplane::{
         DefaultProduceRequest, DefaultPartitionRequest, TopicProduceData, PartitionProduceData,
     },
     api::RequestMessage,
+    ErrorCode,
 };
-use fluvio_controlplane_metadata::partition::Replica;
+use fluvio_controlplane_metadata::{partition::Replica, topic::CompressionAlgorithm};
 use fluvio_future::timer::sleep;
 use fluvio_socket::{MultiplexerSocket, FluvioSocket};
 use flv_util::fixture::ensure_clean_dir;
@@ -40,6 +41,8 @@ async fn test_produce_basic() {
     let topic = "test_produce";
     let test = Replica::new((topic, 0), 5001, vec![5001]);
     let test_id = test.id.clone();
+    ctx.replica_localstore().sync_all(vec![test.clone()]);
+
     let replica = LeaderReplicaState::create(test, ctx.config(), ctx.status_update_owned())
         .await
         .expect("replica");
@@ -114,10 +117,82 @@ async fn test_produce_basic() {
 
     for idx in 0_u16..2_u16 {
         assert_eq!(
+            produce_response.responses[0].partitions[idx as usize].error_code,
+            ErrorCode::None
+        );
+        assert_eq!(
             produce_response.responses[0].partitions[idx as usize].base_offset as u16,
             records_per_request * (3 + idx)
         );
     }
+
+    server_end_event.notify();
+    debug!("terminated controller");
+}
+
+#[fluvio_future::test(ignore)]
+async fn test_produce_invalid_compression() {
+    let test_path = temp_dir().join("test_stream_fetch");
+    ensure_clean_dir(&test_path);
+    let port = portpicker::pick_unused_port().expect("No free ports left");
+
+    let addr = format!("127.0.0.1:{}", port);
+    let mut spu_config = SpuConfig::default();
+    spu_config.log.base_dir = test_path;
+    let ctx = GlobalContext::new_shared_context(spu_config);
+
+    let server_end_event = create_public_server(addr.to_owned(), ctx.clone()).run();
+
+    // wait for stream controller async to start
+    sleep(Duration::from_millis(100)).await;
+
+    let client_socket =
+        MultiplexerSocket::new(FluvioSocket::connect(&addr).await.expect("connect"));
+    let topic = "test_produce";
+    let mut test = Replica::new((topic, 0), 5001, vec![5001]);
+    test.compression_type = CompressionAlgorithm::Gzip;
+    let test_id = test.id.clone();
+    ctx.replica_localstore().sync_all(vec![test.clone()]);
+
+    let replica = LeaderReplicaState::create(test, ctx.config(), ctx.status_update_owned())
+        .await
+        .expect("replica");
+    ctx.leaders_state().insert(test_id, replica.clone()).await;
+
+    // Make three produce requests with <records_per_request> records and check that returned offset is correct
+    let records_per_request = 9;
+    let records = create_filter_records(records_per_request)
+        .try_into()
+        .expect("filter records");
+
+    let mut produce_request = DefaultProduceRequest {
+        ..Default::default()
+    };
+
+    let partition_produce = DefaultPartitionRequest {
+        partition_index: 0,
+        records,
+    };
+    let topic_produce_request = TopicProduceData {
+        name: topic.to_owned(),
+        partitions: vec![partition_produce],
+        ..Default::default()
+    };
+
+    produce_request.topics.push(topic_produce_request);
+
+    let produce_response = client_socket
+        .send_and_receive(RequestMessage::new_request(produce_request))
+        .await
+        .expect("send offset");
+
+    // Check error code
+    assert_eq!(produce_response.responses.len(), 1);
+    assert_eq!(produce_response.responses[0].partitions.len(), 1);
+    assert_eq!(
+        produce_response.responses[0].partitions[0].error_code,
+        ErrorCode::CompressionError
+    );
 
     server_end_event.notify();
     debug!("terminated controller");