+++ crates/fluvio/src/producer/config.rs
@@ -1,9 +1,6 @@
 use std::time::Duration;
 
-use tracing::debug;
-
 use derive_builder::Builder;
-use once_cell::sync::Lazy;
 
 use fluvio_compression::Compression;
 
@@ -11,16 +8,6 @@ use crate::producer::partitioning::{Partitioner, SiphashRoundRobinPartitioner};
 
 const DEFAULT_LINGER_MS: u64 = 100;
 const DEFAULT_BATCH_SIZE_BYTES: usize = 16_384;
-static DEFAULT_COMPRESSION_CODEC: Lazy<Compression> = Lazy::new(|| {
-    use std::env;
-    let var_value =
-        env::var("FLV_CLIENT_DEFAULT_COMPRESSION_CODEC").unwrap_or_else(|_| "none".to_string());
-    let compression: Compression = var_value.parse().unwrap_or_else(|e| {
-        debug!(?e, "unknown compression format in FLV_CLIENT_DEFAULT_COMPRESSION_CODEC, using Compression::None");
-        Compression::None
-    });
-    compression
-});
 
 fn default_batch_size() -> usize {
     DEFAULT_BATCH_SIZE_BYTES
@@ -32,10 +19,6 @@ fn default_linger_duration() -> Duration {
 
 fn default_partitioner() -> Box<dyn Partitioner + Send + Sync> {
     Box::new(SiphashRoundRobinPartitioner::new())
-}
-
-fn default_compression() -> Compression {
-    *DEFAULT_COMPRESSION_CODEC
 }
 
 /// Options used to adjust the behavior of the Producer.
@@ -56,8 +39,10 @@ pub struct TopicProducerConfig {
     pub(crate) partitioner: Box<dyn Partitioner + Send + Sync>,
 
     /// Compression algorithm used by Fluvio producer to compress data.
-    #[builder(default = "default_compression()")]
-    pub(crate) compression: Compression,
+    /// If there is a topic level compression and it is not compatible with this setting, the producer
+    /// initialization will fail.
+    #[builder(setter(into, strip_option), default)]
+    pub(crate) compression: Option<Compression>,
 }
 
 impl Default for TopicProducerConfig {
@@ -66,7 +51,7 @@ impl Default for TopicProducerConfig {
             linger: default_linger_duration(),
             batch_size: default_batch_size(),
             partitioner: default_partitioner(),
-            compression: default_compression(),
+            compression: None,
         }
     }
 }